requirements.txt
Django == 1.11.1
requests == 2.13.0
python_dateutil == 2.6.0
python-memcached==1.58